added settleIt configuration
