--- src/app/Account.ts
@@ -1,5 +1,0 @@
-export default class Account{
-
-
-
-}
