+++ src/app/ReasonRoot.ts
@@ -1,26 +1,19 @@
-//On the
-
-declare const require: any;
-
 type WhichCopy = "original" | "local" | "suggestion";
 
-const hyperHTML = require('hyperhtml');
-//const firebase = require('firebase');
-
+import hyperHTML = require('hyperhtml');
 import Root from './Root';
 import Dict from './Dict';
 import SettleIt from'./SettleIt';
 import Score from './Score';
 import Claim from './Claim';
 import Firebase from './Firebase';
-import Display from './Display';
 
 export default class ReasonRoot {
     userName: string = 'Sign In';
     rrRef: any;//The current firebase reference to the ReasonRoot object
     scores: Dict<Score>;
     claims: Dict<Claim>;
-    settleIt: SettleIt;node
+    settleIt: SettleIt;
     mainScore: Score;
     render: any;
     settings: any = {};
@@ -30,22 +23,24 @@ export default class ReasonRoot {
     db: any;
     rr: Root = new Root();
     whichCopy: WhichCopy;
-    //settingsVisible: boolean = false;
+    settingsVisible: boolean = false;
     listenerRefs: any[] = new Array<any>();
     canWrite: boolean;
     mainId: any;
-    firebase: Firebase = new Firebase();
+    firebase: Firebase;
     score: Score = new Score();
     claim: Claim = new Claim();
-    display: Display;
 
-    constructor(claimElement?: Element) {
+    constructor(firebase: Firebase, claimElement?: Element) {
       if(claimElement){
+        //this.render is a pointer to the Claim HTML tag.
         this.render = hyperHTML.bind(claimElement);
         this.settleIt = new SettleIt();
         this.rr = JSON.parse(claimElement.getAttribute('root'));
-        this.firebase.firebaseInit(this.rr, this.canWrite);
+        this.firebase = firebase;
+        //this.firebase.onAuthStateChanged(this.rr, this.canWrite);
         this.changeWhichCopy("original");
+
         //this.attachDB();
         //this.initRr();
         //this.update();
@@ -58,9 +53,11 @@ export default class ReasonRoot {
         this.scores = this.createDict(this.claims);
         this.mainScore = this.scores[this.rr.mainId];
         this.mainScore.isMain = true;
-        this.display = new Display(this.render, this.settings);
-        this.settleIt.calculate(this.rr.mainId, this.claims, this.scores)
-        this.setDisplayState(this.selectedScore);
+
+        // Returns a Object loaded with the result of global variables: mainId, claims, scores
+        // TODO Why calculate an this.settleIt.calculate is called at the same time?
+        //this.settleIt.calculate(this.rr.mainId, this.claims, this.scores);
+        this.setDisplayState();
         this.calculate();
     }
 
@@ -80,13 +77,13 @@ export default class ReasonRoot {
 
             // TODO localStorage is null
             let rr = localStorage.getItem(this.savePrefix + this.rr.mainId);
-            console.log(rr);
             if (rr) {
                 this.rr = JSON.parse(rr);
             }
         } else {
+          // Original and suggestion both mean remote
           // This is the problem
-            this.firebase.firebaseInit(this.rr, this.canWrite);
+            //this.firebase.firebaseInit(this.rr, this.canWrite);
             if (whichCopy === "original") {
                 this.rrRef = this.firebase.db.ref('roots/' + this.rr.mainId);
             }
@@ -94,13 +91,13 @@ export default class ReasonRoot {
                 //to do Find the ID of my suggestion
                 this.rrRef = this.firebase.db.ref('roots/' + this.rr.mainId);
             }
-            //this.attachDB();
+            this.attachDB();
         }
-
         this.initRr();
-        this.display.update(this.renderNode(this.scores[this.rr.mainId]));
+        this.update();
     }
 
+    //This may be on Firebase and it sync reason root remote data with local one
     attachDB() {
       let claimsRef = this.rrRef.child('claims');
       this.listenerRefs.push(claimsRef);
@@ -120,36 +117,140 @@ export default class ReasonRoot {
       }
     }
 
+    //This may be on Firebase
     claimsFromDB(data: any) {
+        // Here claims are pulled from firebase.
+        // value is the reason root object
         let value = data.val();
         if (value) {
             this.rr.claims = value;
             this.claims = value;
             this.calculate();
-            this.display.update(this.renderNode(this.scores[this.rr.mainId]));
+            this.update();
         }
     }
 
+    //This may be on Firebase
     claimFromDB(data: any) {
         let value = data.val();
         if (value) {
             let claim: Claim = value;
             this.claims[claim.claimId] = claim;
             this.calculate();
-            this.display.update(this.renderNode(this.scores[this.rr.mainId]));
+            this.update();
         }
     }
 
-    setDisplayState(selectedScore?: Score): void {
-        this.display.clearDisplayState(this.scores);
-        this.mainScore = this.display.setDisplayStateLoop(this.mainScore, this.claims, this.scores, this.selectedScore);
+    setDisplayState(): void {
+        this.clearDisplayState();
+        this.setDisplayStateLoop(this.mainScore);
     }
 
+    clearDisplayState(): void {
+        for (let scoreId in this.scores) {
+            if (this.scores.hasOwnProperty(scoreId)) {
+                this.scores[scoreId].displayState = "notSelected";
+            }
+        }
+    }
+
+    //Set state of claims every time selected score change
+    setDisplayStateLoop(score: Score): void {
+        // Compare the param score whit selected Score if they are the same set score state to "selected".
+        if (score == this.selectedScore)
+            score.displayState = "selected";
+        let childIds = this.claims[score.claimId].childIds;
+        if(childIds !== undefined){
+          for (let childId of childIds) {
+              let childScore = this.scores[childId];
+              //process the children first/
+              this.setDisplayStateLoop(childScore);
+
+              if (childScore == this.selectedScore) {
+                  score.displayState = "parent";
+                  //Set Siblings
+                  for (let siblingId of this.claims[score.claimId].childIds) {
+                      let siblingScore = this.scores[siblingId];
+                      if (siblingScore.displayState != "selected")
+                          siblingScore.displayState = "sibling";
+                  }
+              }
+
+              if (childScore.displayState == "ancestor" || childScore.displayState == "parent")
+                  score.displayState = "ancestor";
+
+              if (score == this.selectedScore)
+                  childScore.displayState = "child";
+          }
+        }
+    }
+
+    // This displays the settngs menu of the main claim.
+    update(): void {
+
+      // if (!this.settings.noAutoSave)
+      //     localStorage.setItem(this.savePrefix + this.root.mainId, JSON.stringify(this.scores));
+
+      this.render`
+      <div class="${'rr' +
+          (this.settings.hideScore ? ' hideScore' : '') +
+          (this.settings.hidePoints ? ' hidePoints' : '') +
+          (this.settings.hideClaimMenu ? ' hideClaimMenu' : '') +
+          (this.settings.hideChildIndicator ? ' hideChildIndicator' : '') +
+          (this.settings.showSiblings ? ' showSiblings' : '') +
+          (this.settings.showCompetition ? ' showCompetition' : '')
+
+          }">
+          <div class = "${'settingsHider ' + (this.settingsVisible ? 'open' : '')}">
+              <input type="checkbox" id="hideScore" bind="hideScore" value="hideScore" onclick="${this.updateSettings.bind(this, this.settings)}">
+              <label for="hideScore">Hide Score</label>
+              <input type="checkbox" id="hidePoints" bind="hidePoints" value="hidePoints" onclick="${this.updateSettings.bind(this, this.settings)}">
+              <label for="hidePoints">Hide Points</label>
+              <input type="checkbox" id="noAutoSave" bind="noAutoSave" value="noAutoSave" onclick="${this.updateSettings.bind(this, this.settings)}">
+              <label for="noAutoSave">No Auto Save</label>
+              <input type="checkbox" id="showSiblings" bind="showSiblings" value="showSiblings" onclick="${this.updateSettings.bind(this, this.settings)}">
+              <label for="showSiblings">Show Sibllings</label>
+              <input type="checkbox" id="hideClaimMenu" bind="hideClaimMenu" value="hideClaimMenu" onclick="${this.updateSettings.bind(this, this.settings)}">
+              <label for="hideClaimMenu">Hide Claim Menu</label>
+              <input type="checkbox" id="hideChildIndicator" bind="hideChildIndicator" value="hideChildIndicator" onclick="${this.updateSettings.bind(this, this.settings)}">
+              <label for="hideChildIndicator">Hide Child Indicator</label>
+              <input type="checkbox" id="showCompetition" bind="showCompetition" value="showCompetition" onclick="${this.updateSettings.bind(this, this.settings)}">
+              <label for="showCompetition">Show Competition</label>
+
+              <input value="${this.replaceAll(JSON.stringify(this.rr), '\'', '&#39;')}"></input>
+
+              <div  onclick="${this.firebase.SignIn.bind(this)}">
+                      [${this.userName} ]
+              </div>
+         </div>
+          <div>${this.renderNode(this.scores[this.rr.mainId])}</div>
+          <div class="settingsButton" onclick="${this.toggleSettings.bind(this)}">
+              ⚙
+          </div>
+      </div>`;
+    }
+
+    // Settings is part of the Reason Root's main claim, this should be on Reason Root component
+    updateSettings(settings: any, event: any): void {
+        settings[event.srcElement.getAttribute("bind")] = event.srcElement.checked;
+        this.update();
+        if (event) event.stopPropagation();
+    }
+
+    toggleSettings(event: Event, node: any): void {
+        this.settingsVisible = !this.settingsVisible;
+        this.update();
+    }
+
+    replaceAll(target: string, search: string, replacement: string): string {
+        return target.split(search).join(replacement);
+    }
+
+    // This render a unique Claim, this should be part of Claim component,both main claim and child climes use it.
     renderNode(score: Score, parent?: Score): void {
       var claim: Claim = this.claims[score.claimId];
-      var claims = this.claims;
       var wire = hyperHTML.wire(score);
-      if (this.display.animateNumbers(this.scores)) setTimeout(() => this.display.update(this.renderNode(this.scores[this.rr.mainId])), 100);
+      this.animateNumbers();
 
       var result = wire`
               <li id="${claim.claimId}" class="${
@@ -235,11 +336,28 @@ export default class ReasonRoot {
       return result;
     }
 
+    //Check for animating numbers of the Reason Root's main claim, this should be on Reason Root component
+    animateNumbers() {
+        var found = false;
+        for (var scoreId in this.scores) {
+            var s = this.scores[scoreId];
+            if (s.weightedPercentage != s.animatedWeightedPercentage) {
+                found = true;
+                var difference = s.weightedPercentage - s.animatedWeightedPercentage
+                if (Math.abs(difference) < .01)
+                    s.animatedWeightedPercentage = s.weightedPercentage
+                else
+                    s.animatedWeightedPercentage += difference / 100;
+            }
+        }
+        if (found) setTimeout(() => this.update(), 100);
+    }
+
     selectScore(score: Score, e: Event): void {
       if (score != this.selectedScore) {
           this.selectedScore = score;
           this.setDisplayState();
-          this.display.update(this.renderNode(this.scores[this.rr.mainId]));
+          this.update();
       }
     };
 
@@ -249,42 +367,29 @@ export default class ReasonRoot {
 
 
     calculate(): void {
-      this.settleIt.calculate(this.rr.mainId, this.claims, this.scores)
+      this.settleIt.calculate(this.rr.mainId, this.claims, this.scores);
     };
 
     // The logic of this functionalities: addClaim, updateClaim, and removeClaim were moved
     // to their own class file, and then they only should be called from other classes like this.
 
-    createReasonRoot(content, citation) {
-      let database = this.firebase.getDatabase();
-      let mainId = this.claim.newId();
-      let currentUserId = this.firebase.getCurrentUser().uid;
-
-      let permission = {};
-      permission[mainId] = true;
-      database.ref('permissions/user/' + currentUserId).update(permission);
-
-      let mainClaim = new Claim(mainId, true);
-      mainClaim.citation = citation;
-      mainClaim.content = content;
-
-      let claim = {};
-      claim[mainId] = Object.assign({}, mainClaim);
-
-      database.ref('roots/' + mainId).set({
-        claims: claim,
-        mainId: mainId
-      });
-
-      let data = this.firebase.getDataById(mainId);
-      var element = document.createElement("claim");
-      element.setAttribute("root", data);
-      document.body.appendChild(element);
+    createReasonRoot() {
+      let claimId = this.firebase.createReasonRoot();
+      this.appendReasonRoot(claimId);
     };
 
-    isEmpty(obj) {
-    return Object.keys(obj).length === 0;
-    }
+    appendReasonRoot(mainId){
+      let element = document.createElement("claim");
+      this.firebase.getDataById(mainId)
+        .then(data => {
+          element.setAttribute("root", data);
+        })
+       .then(() => {
+         document.body.appendChild(element);
+       }
+     );
+    };
+
 
     addClaim(isProMain: boolean, event?: Event) {
       let childClaim: Claim;
@@ -292,8 +397,8 @@ export default class ReasonRoot {
       childClaim = this.claim.add(this.selectedScore, isProMain, this.scores, this.claims);
       this.firebase.addData(this.rr, parentClaim, childClaim);
       this.calculate();
-      this.setDisplayState(this.selectedScore);
-      this.display.update(this.renderNode(this.scores[this.rr.mainId]));
+      this.setDisplayState();
+      this.update();
 
       if (event) event.stopPropagation();
     };
@@ -303,21 +408,23 @@ export default class ReasonRoot {
       this.firebase.updateData(this.rr, claim);
       //update the UI
       this.calculate();
-      this.display.update(this.renderNode(this.scores[this.rr.mainId]));
+      this.update();
 
     };
 
     removeClaim(claim: Claim, parentScore: Score, event: Event): void {
+      let parentClaim: Claim = this.claims[parentScore.claimId];
       this.claim.remove(claim, this.claims, parentScore, event);
-      this.firebase.deleteData(this.rr, claim);
+      this.firebase.deleteData(this.rr, parentClaim, claim);
       this.calculate();
-      this.setDisplayState(this.selectedScore);
-      this.display.update(this.renderNode(this.scores[this.rr.mainId]));
+      this.setDisplayState();
+      this.update();
     };
 
     editClaim(score: Score, event?: Event): void {
+
       this.settings.isEditing = !this.settings.isEditing;
-      this.display.update(this.renderNode(this.scores[this.rr.mainId]));
+      this.update();
       if (event) event.stopPropagation();
     }
